+++ install/config/all.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-run_phase "$HANDAAN_INSTALL/config/shell.sh"
+run_phase "$HANDAAN_INSTALL/config/home-dotfiles.sh"
 run_phase "$HANDAAN_INSTALL/config/env-bootstrap.sh"
 run_phase "$HANDAAN_INSTALL/config/seed-config.sh"
 run_phase "$HANDAAN_INSTALL/config/locale.sh"